requirements.txt
eth_account==0.5.6
Pillow==8.4.0
pycryptodome==3.11.0
pyperclip==1.8.2
qrcode==7.3.1
